requirements.txt
matplotlib
ipython
Pillow
